# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sneaker Marketplace API",
    "description": "Peer-to-peer sneaker buying and selling platform. Use the Authorize button to enter your JWT token (format: `Bearer <token>`).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "API base path" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id":       { "type": "integer" },
          "username": { "type": "string" },
          "email":    { "type": "string" },
          "role":     { "type": "string", "enum": ["buyer","seller","admin"] },
          "token":    { "type": "string" }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id":         { "type": "integer" },
          "buyerId":    { "type": "integer" },
          "totalPrice": { "type": "number" },
          "status":     { "type": "string", "enum": ["pending","confirmed","cancelled","completed"] },
          "createdAt":  { "type": "string", "format": "date-time" },
          "orderItems": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id":        { "type": "integer" },
                "sneakerId": { "type": "integer" },
                "quantity":  { "type": "integer" },
                "unitPrice": { "type": "number" }
              }
            }
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id":          { "type": "integer" },
          "name":        { "type": "string" },
          "description": { "type": "string" },
          "createdAt":   { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username","email","password"],
                "properties": {
                  "username": { "type": "string", "example": "sneakerhead99" },
                  "email":    { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "Password123!" },
                  "role":     { "type": "string", "enum": ["buyer","seller","admin"], "example": "buyer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Account created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Email or username already exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log in and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email","password"],
                "properties": {
                  "email":    { "type": "string", "example": "seller1@example.com" },
                  "password": { "type": "string", "example": "Seller123!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Bad Request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Place a new order",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sneakerIds"],
                "properties": {
                  "sneakerIds": { "type": "array", "items": { "type": "integer" }, "example": [1] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Sneaker not found" },
          "409": { "description": "Sneaker already sold" }
        }
      },
      "get": {
        "tags": ["Orders"],
        "summary": "Get orders (scoped by role)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending","confirmed","cancelled","completed"] } }
        ],
        "responses": {
          "200": { "description": "List of orders", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } } } } },
          "400": { "description": "Invalid status" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get an order by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "responses": {
          "200": { "description": "Order found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
          "400": { "description": "Invalid ID" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Order not found" }
        }
      },
      "put": {
        "tags": ["Orders"],
        "summary": "Update order status",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["pending","confirmed","cancelled","completed"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Order not found" },
          "409": { "description": "Already cancelled or completed" }
        }
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Delete an order",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "responses": {
          "200": { "description": "Order deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Order not found" },
          "409": { "description": "Already confirmed or completed" }
        }
      }
    },
    "/categories": {
      "post": {
        "tags": ["Categories"],
        "summary": "Create a new category",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name":        { "type": "string", "example": "Skate" },
                  "description": { "type": "string", "example": "Skateboarding shoes" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Category created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden: Not an admin" },
          "409": { "description": "Name already exists" }
        }
      },
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of categories", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } } } } },
          "400": { "description": "Invalid query parameter" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get a category by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "responses": {
          "200": { "description": "Category found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "400": { "description": "Invalid ID" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Category not found" }
        }
      },
      "put": {
        "tags": ["Categories"],
        "summary": "Update a category",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name":        { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Category updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden: Not an admin" },
          "404": { "description": "Category not found" },
          "409": { "description": "Name conflict" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete a category",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "responses": {
          "200": { "description": "Category deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden: Not an admin" },
          "404": { "description": "Category not found" },
          "409": { "description": "Has sneaker listings assigned" }
        }
      }
    }
  }
}